# docs/swagger.yaml
definitions:
...
  controllers.phoneInput:
    properties:
      Cameraback:
        type: string
      Camerafront:
        type: string
      battery:
        type: string
      brandID:
        type: integer
      chipset:
        type: string
      display:
        type: string
      name:
        type: string
      pict:
        type: string
      ram:
        type: string
      username:
        type: string
    type: object
  controllers.phoneUser:
    properties:
      username:
        type: string
    type: object
  controllers.reviewInput:
    properties:
      isi:
        type: string
      phoneID:
        type: integer
      userID:
        type: integer
    type: object
...
  models.Phone:
    properties:
      Cameraback:
        type: string
      Camerafront:
        type: string
      battery:
        type: string
      brandID:
        type: integer
      chipset:
        type: string
      created_at:
        type: string
      display:
        type: string
      id:
        type: integer
      name:
        type: string
      pict:
        type: string
      ram:
        type: string
    type: object
  models.Review:
    properties:
      created_at:
        type: string
      id:
        type: integer
      isi:
        type: string
      phoneID:
        type: integer
      userID:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /phone/:
    post:
      description: Creating a new Phone (only username with role id 1 can do this).
      parameters:
      - description: the body to create a new Phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - ApiKeyAuth: []
      summary: Create New Phone.
      tags:
      - Phone
  /phone/{id}:
    delete:
      description: Delete a Phone by id (only username with role id 1 can do this).
      parameters:
      - description: the User
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneUser'
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Phone.
      tags:
      - Phone
    get:
      description: Get a Phone by id.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      summary: Get Phone.
      tags:
      - Phone
    patch:
      description: Update a Phone by id (only username with role id 1 can do this).
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - ApiKeyAuth: []
      summary: Update Phone.
      tags:
      - Phone
  /phone/brand/{id}:
    get:
      description: Get a Phone by Brand id.
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      summary: Get Phone By Brand.
      tags:
      - Phone
  /phone/picture/{id}:
    get:
      description: Get a Phone Picture.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: ""
      summary: Get Phone Picture.
      tags:
      - Phone
  /phone/upload/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Upload Phone Picture (.png only!).
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - description: image (.png only)
        in: formData
        name: file
        required: true
        type: file
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Upload Picture Phone.
      tags:
      - Phone
...
  /review/:
    post:
      description: Creating a new Review.
      parameters:
      - description: the body to create a new Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - ApiKeyAuth: []
      summary: Create New Review.
      tags:
      - Review
  /review/{id}:
    delete:
      description: Delete a Review by id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Review.
      tags:
      - Review
    get:
      description: Get a Review by id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Get Review.
      tags:
      - Review
    patch:
      description: Update Review by id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - ApiKeyAuth: []
      summary: Update Review.
      tags:
      - Review
  /review/phone/{id}:
    get:
      description: Get a Review by Phone id.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Get Review by Phone id.
      tags:
      - Review
...
swagger: "2.0"